ai-dispatch-core/requirements.txt
pandas
Flask
python-dotenv
python-telegram-bot
